Refactor code, separating classes in different files
